refactor: extract ApplicationController responsibilities into concerns
ApplicationController was accumulating too many responsibilities in a single
269-line file. Extract four coherent groups into ActiveSupport::Concern modules
so each file has a single clear purpose and is easy to find and modify in
isolation. No behaviour changes — before_action and rescue_from declarations
remain in ApplicationController in their original order.

Concerns extracted:
- CacheControl: set_cache and the three response headers it sets
- FlagsmithSetup: anonymous identity cookie creation and Current.flagsmith_identity assignment
- BotProtection: X-Robots-Tag suppression for historical search dates
- ErrorHandling: all rescue_from handlers and their supporting private methods
