+++ app/controllers/concerns/bot_protection.rb
@@ -1,0 +1,13 @@
+module BotProtection
+  extend ActiveSupport::Concern
+
+  private
+
+  def bots_no_index_if_historical
+    return if @search.today?
+
+    response.headers['X-Robots-Tag'] = 'none'
+  rescue StandardError
+    response.headers['X-Robots-Tag'] = 'none'
+  end
+end
